style: remove unnecessary cast from byte to int

+++ src/org/jcodings/util/BytesHash.java
@@ -62,7 +62,7 @@ public final class BytesHash<V> extends Hash<V>{
 
     public static int hashCode(byte[]bytes, int p, int end) {
         int key = 0;
-        while (p < end) key = ((key << 16) + (key << 6) - key) + (int)(bytes[p++]); // & 0xff ? we have to match jruby string hash
+        while (p < end) key = ((key << 16) + (key << 6) - key) + bytes[p++]; // & 0xff ? we have to match jruby string hash
         key = key + (key >> 5);
         return key;
     }
